refactor: 상품 주문 완료 시점에 장바구니 업데이트 및 쿠폰/할인/체크 상태 초기화가 실행되도록 수정
