+++ src/hooks/index.ts
@@ -5,3 +5,4 @@ export { default as useDiscountCalculator } from './useDiscountCalculator';
 export { default as useCoupons } from './useCoupons';
 export { default as useApplyCoupons } from './useApplyCoupons';
 export { default as useCouponApplyButton } from './useCouponApplyButton';
+export { default as useClearOrderedItems } from './useClearOrderedItems';
